Improve the test regarding the required argument of addAction()
The addAction() method requires either a function or a table that has
an implementation of the call() metamethod.  This patch improves the
unit test for the arguments of addAction() by calling it with a table
that does not have that metamethod.  In this situation the method
should always result in an error, which is what the test expects.