Improve the unit test for giving addAction() a table
We define a table within the test function itself.  This patch deletes
that table and creates a new one in the same scope as the other
actions we use within that block of tests.  This will allow us to
write more tests in that describe() section which use tables instead
of functions for actions.