Refactor the test for valid actions into a private function
The method addAction() first determines if the argument it receives is
a valid action or not.  This patch moves that test into a separate
function which is private to the Luvent module.  This not only
improves readability but also gives the means, if necessary, to test
other arguments and data to see if they are valid actions.