Return a table index when searching for actions
We can use callsAction() to determine whether or not an event will use
an action or not.  However, given an action we have no way to find
where it appears in the ‘actions’ table of an event; the best we can
do is manually loop through the table until we find the action and
then return its index.  We already perform this exact type of loop in
the callsAction() method.

So this patch extracts the logic of callsAction() into a separate,
private function that we name findAction().  This function will tell
us if an event has an action, but on top of that it will return the
index for that action if it exists.  We can use that index to improve
the code of other methods, and with this patch we rewrite
callsAction() to simply call findAction() and only return the first
value, preserving the previous behavior.