Fix the definition of isValidAction()
The document claims that the function returns a boolean but this is
not what the code does.  This patch changes the code so that it
actually implements the behavior described in the documentation.

The patch also makes addAction() raise an error if given an argument
for which isValidAction() is false.  This change maintains the
existing behavior.