Give all actions an ID
This will allow us to find actions by something besides the value of
their ‘callable’ property.  As the source comment describes, that will
be useful to do things such as remove actions that we created by
giving the addAction() method an anonymous function; we would not have
access to the anonymous function to give to removeAction() later, but
we could teach removeAction() to accept IDs as well.

This patch also redefines the equality test for actions so that it
compares their IDs, which semantically makes more sense to me.