Teach findAction() to accept action IDs
This patch makes it possible for us to find actions in an event using
nothing but the ID returned by methods like addAction().  As a
side-effect we can now remove actions using only the ID since the
implementation for removal relies on findAction() to figure out which
action to get rid of.

With this patch in place all current tests will succeed.