Remove the private function addActionToEvent()
The function is no longer necessary now that we have addAction() as
the sole public method for adding actions to events.  Previously we
needed this function to help with addActionWithInterval() but now
that the method is no longer part of the API we simply the code base
by getting rid of this internal function.