Rename callsAction() to hasAction()
Now that it is possible to disable an action it no longer makes sense
to use the name callsAction(), since the method only tests whether or
not the event has knowledge of the action.  That means callsAction()
returns true for actions that the event will not actually call,
i.e. disabled actions.  It makes more sense to simply ask if the event
‘has’ an action or not.