Improve the unit tests for action IDs
This patch adds one new test which ensures that Luvent creates the
same ID for an action when adding that action to separate events.
Then the patch corrects an existing test; one test checks to see if an
action ID is a string but we should not do this as it can give the
wrong impression that IDs will always be strings, which is not true.
So the patch modifies the test to assert that the action ID is a
‘truthy’ value, i.e. anything Lua would consider true in a boolean
context, which is all the precision we actually need for that test.