Implement setting and removing action priorities
This patch provides two methods for setting and removing the
priorities of actions.  In order for the priorities to affect the
behavior of events we must sort all actions.  This patch performs that
sort during calls to the trigger() method just before it starts
invoking actions.  This is admittedly inefficient; often we expect we
will trigger an event without having changed the priorities of any
actions and that makes the sort wasted time.  We could perform the
sort any time we do something that directly affects an event’s table
of actions, e.g. adding or removing one, but for now we will perform
the sort in one place for the sake of simplicity.