README: Start reorganizing the API documentation
This patch starts the process of breaking down the API documentation
into chunks of different examples, each showcasing a different part of
the library.  This patch also adds an example for enabling and
disabling actions and how that compares to simply removing actions.

The full list of methods includes ‘removeActionInterval()’, which does
not currently exist.  That is an oversight that a future patch will
address by implementing it.