Do not test the default order of actions without priorities
The README says that Luvent does not guarantee the order in which it
will invoke actions which have the same priority, and by default all
actions have the same priority.  However, the test suite contains a
test to make sure that Luvent, by default, executes actions in the
order we add them to the event.  This contradicts the README’s
description of Luvent’s behavior and the intended design.  This patch
removes the unit test so that it will not confuse users over whether
Luvent guarantees that order of execution or not.