Allow modification of the Luvent and Luvent.Action metatables
Originally Luvent did not allow changing the metatables of its two
classes as a means to provide some amount of stability.  However,
after reading this blog post…

    http://kiki.to/blog/2014/04/04/rule-3-allow-monkeypatching/

…I was persuaded that it is more beneficial for developers to allow
changing the metatables.  Some games may want to implement custom
behavior in the metatables and Luvent should be extensible in that
way, because ease-of-implementation will boost overall community
adoption and support.