Makefile: Update ‘make tests’ to account for a problem with busted
I upgraded to the latest release candidate of busted 2.0 and now the
program no longer reads settings from the ‘.busted’ file.  So this
patch changes the Makefile so that it will run the test suite, which
requires us to pass some arguments to busted.  Previously we would
just run busted and let it read the equivalent of those arguments from
the ‘.busted’ file, but since that no longer works we have to do this.