Correct a unit test for setActionInterval()
One test for setActionInterval() tells an event not to invoke actions
more than once per second.  We then trigger that event in a continuous
‘while (true)’ loop for three seconds.  The expectation is for the
event to only trigger actions three times during that period.
However, because we run that loop for exactly three seconds, and in an
unexact way, the test will sometimes pass and sometimes fail.  We’re
cutting things a little too close with our timing.  So this patch
extends the loop to 3.5 seconds.  The extra time gets rid of the race
to finish first between trigger() and the loop.