[Style] Rename a variable in createActionSetter() to squash a warning
Running Luachuck on the source code brings up an error about how we
reuse the 'newValue' variable within createActionSetter().  This patch
changes the variable name to squash that error.  It has no effect on
the logic or behavior of the code, but is simply a clean-up so that
Luacheck reports no warnings or errors.